Add producer and recipient modules
