+++ crates/core/tedge_actors/src/lib.rs
@@ -1,11 +1,15 @@
 mod actor;
 mod error;
 mod message;
+mod producer;
+mod recipient;
 mod runtime;
 
 pub use actor::*;
 pub use error::*;
 pub use message::*;
+pub use producer::*;
+pub use recipient::*;
 pub use runtime::*;
 
 #[cfg(test)]
